Moved club and event details inside the dashboard route
